refactoring cod. added .env
