--- src/.env
@@ -1,3 +1,0 @@
-# Keys for API
-REACT_APP_BASE_URL=https://social-network.samuraijs.com/api/1.1/
-REACT_APP_API_KEY=46bf5cab-c958-45f9-89c4-0ee6d1b7ed40
