+++ src/features/ErrorPage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { isRouteErrorResponse, useRouteError, useNavigate } from "react-router-dom";
 import { Button, makeStyles, Typography } from "@material-ui/core";
 import errorImg from "../assets/images/error.jpg";
-import { theme } from "../styles/general";
+import { theme } from "../styles/common";
 
 const useStyles = makeStyles({
   container: {
