+++ src/hooks/useAppDispatch.tsx
@@ -1,4 +1,8 @@
+import { useSelector } from "react-redux";
 import {useDispatch} from "react-redux";
-import {AppRootDispatch} from "../app/store";
+import {AppRootDispatch, AppRootState} from "../app/store";
 
 export const useAppDispatch = () => useDispatch<AppRootDispatch>()
+export const useAppSelector = () => useSelector<AppRootState>(state => state)
+
+
