+++ src/styles/common.ts
